Seperate Op into interface and abstract class
